Going to refactor to app-of-apps pattern.
