+++ atils/pyproject.toml
@@ -16,12 +16,15 @@ deepdiff = "^6.3.0"
 shutils = "^0.1.0"
 requests = "^2.29.0"
 hvac = "^1.1.0"
+dynaconf = "^3.1.12"
 
 [tool.poetry.group.dev.dependencies]
 MonkeyType = "22.2.0"
 black = "22.12.0"
 pytest = "7.2.0"
 
+[tool.poetry.scripts]
+atils = "atils.atils:main"
 
 [build-system]
 requires = ["poetry-core"]
