+++ mac-setup/Brewfile
@@ -43,3 +43,4 @@ brew 'helm'
 brew 'rke'
 brew 'yq'
 brew 'vault'
+brew 'pipx'
